+++ drivers/soc/eswin/ai_driver/npu/event.c
@@ -130,7 +130,7 @@ void dla_event_sink_dump_config(struct dla_processor_group *group)
 }
 
 int dla_event_sink_prepare_prog_data(struct win_executor *executor, int rdma,
-				int tensor_idx, u16 op_idx,
+				int tensor_idx, u32 op_idx,
 				union dla_operation_container *operation_desc,
 				union dla_surface_container *surface_desc)
 {
@@ -176,7 +176,7 @@ void dla_event_source_dump_config(struct dla_processor_group *group)
 }
 
 int dla_event_source_prepare_prog_data(struct win_executor *executor, int rdma,
-				int tensor_idx, u16 op_idx,
+				int tensor_idx, u32 op_idx,
 				union dla_operation_container *operation_desc,
 				union dla_surface_container *surface_desc)
 {
