+++ drivers/soc/eswin/ai_driver/npu/pdp.c
@@ -580,7 +580,7 @@ exit:
 }
 
 int dla_pdp_prepare_prog_data(struct win_executor *executor, int rdma,
-			      int tensor_idx, u16 op_idx,
+			      int tensor_idx, u32 op_idx,
 			      union dla_operation_container *operation_desc,
 			      union dla_surface_container *surface_desc)
